% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_regression_output.R
\name{AllRegressionOutput}
\alias{AllRegressionOutput}
\title{All Linear Regression List of Returns vs. Risk factors}
\usage{
AllRegressionOutput(risk.factors, returns, kStartingIndex = 2, alpha = 0.02,
  sum.reg.fun)
}
\arguments{
\item{risk.factors}{data frame of risk factors for all companies}

\item{returns}{data frame of return records for all companies}

\item{sum.reg.fun}{function name for specific summary regression}

\item{kStaringIndex}{an integer index}

\item{alfa}{float number to indicate the significant level of a risk factor}
}
\value{
data frame of regression results for all companies
}
\description{
list all regressions from a starting company index
    data for returns and risk factors need be free of NAs.
}
\examples{
summary.regression <- AllRegressionOutput(risk.factors, returns, 2, 0.05, SummaryRegression)

}
