% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_missing_value_returns.R
\name{NoMissingValueReturns}
\alias{NoMissingValueReturns}
\title{Filter out companies with imcompletle records}
\usage{
NoMissingValueReturns(returns)
}
\arguments{
\item{returns}{the variable name for a uploaded csv file of company returns. 
It may has missing values and can have many company returns 
on one file.}
}
\value{
data frame of company returns without missing values
}
\description{
This function allows you delete and filter out columns with missing values.
}
\examples{
nmv.returns <- NoMissingValueReturns(returns)
}
