% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_regression.R
\name{SingleRegression}
\alias{SingleRegression}
\title{Linear Regression of Returns vs. Risk factors}
\usage{
SingleRegression(risk.factors, returns, kCompanyIndex)
}
\arguments{
\item{risk.factors}{data frame of risk factors for all companies}

\item{returns}{data frame of return records for all companies}

\item{kCompanyIndex}{an integer index for a company}
}
\value{
linear regression model of a fit
}
\description{
run regression for a given company index
    data for returns and risk factors need be free of NAs.
}
\examples{
sr <- SingleRegression(risk.factors, returns, kCompanyIndex)

}
