% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_regression.R
\name{SummaryRegression}
\alias{SummaryRegression}
\title{Linear Regression of Returns vs. Risk factors}
\usage{
SummaryRegression(risk.factors, returns, kCompanyIndex, alpha = 0.05)
}
\arguments{
\item{risk.factors}{data frame of risk factors for all companies}

\item{returns}{data frame of return records for all companies}

\item{kCompanyIndex}{an integer index for a company}

\item{alfa}{float number to indicate the significant level of a risk factor}
}
\value{
data frame of one-row regression results
}
\description{
run regression for a given company index
    data for returns and risk factors need be free of NAs.
}
\examples{
sr.df <- SummaryRegression(risk.factors, returns, kCompanyIndex, 0.05)

}
